Intrainz Web Intern Minor Project
Calculator Project

Developing a Calculator after learning basic of HTML, CSS & JavaScript is one of the best ways to understand the core concept and implement it to make something useful.
This Calculator you can modify as per your need to perform both simple and complex calculations. It should include +, -, *, / operations along with calculating modulo (%) and performing square of a number. Also, add AC (All Clear) button to clear the current text and Equal key for performing the operation. The calculator should perform operation using both Decimal and Whole numbers.
You can add more functionality as per your need but the above-mentioned one are a must to have.

Must use the following for development:
 HTML
 CSS(Flexbox or Grid) and Bootstrap
 JavaScript( Event listeners, multiple functions, loops, if/else, InnerText/InnerHtml)
 Use comments wherever possible to make it readable for others.
 Deploy your code to GitHub and merge with Netlify to make it live.
 Do not use any external library for this.
 Build your own logic do not copy paste from google.
 Design can be as per your vision and make it as creative as you can.

The Project will be judged on the Functionality, if all the mentioned things are used or
not, how clean the code is and lastly the design.